# openapi.yaml
openapi: 3.0.3
info:
  title: Cratebase API
  version: "0.1.0"
  description: >
    Dynamic collections, auth, file storage, and realtime over one REST API.
    Error responses always use the shape in `ErrorBody`. Bearer tokens
    (admin or auth-collection) go in `Authorization: Bearer <token>`.
servers:
  - url: /api
security:
  - bearerAuth: []
  - {}
paths:
...
  /collections:
    get:
      summary: List collections (superuser only)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Collection" }
        "401": { $ref: "#/components/responses/Error" }
    post:
      summary: Create a collection (superuser only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CollectionInput" }
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Collection" }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
  /collections/{idOrName}:
    parameters:
      - name: idOrName
        in: path
        required: true
        schema: { type: string }
    get:
      summary: Get a collection (superuser only)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Collection" }
        "404": { $ref: "#/components/responses/Error" }
    patch:
      summary: Update a collection's schema/rules (superuser only)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CollectionInput" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Collection" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      summary: Delete a collection and every record in it (superuser only)
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/Error" }
...
  /files/{collection}/{recordId}/{filename}:
    parameters:
      - name: collection
        in: path
        required: true
        schema: { type: string }
      - name: recordId
        in: path
        required: true
        schema: { type: string }
      - name: filename
        in: path
        required: true
        schema: { type: string }
    get:
      summary: Download a file (gated by the collection's viewRule)
      responses:
        "200":
          description: OK
          content:
            application/octet-stream: { schema: { type: string, format: binary } }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
  schemas:
    ErrorBody:
      type: object
      properties:
        code: { type: integer }
        message: { type: string }
        data:
          type: object
          additionalProperties:
            type: object
            properties:
              code: { type: string }
              message: { type: string }
          description: Per-field validation errors, present when code is 400
...
    FieldSchema:
      type: object
      required: [id, name, type]
      properties:
        id: { type: string }
        name: { type: string }
        type:
          type: string
          enum: [text, editor, number, bool, email, url, date, select, json, relation, file, password]
        required: { type: boolean }
        unique: { type: boolean }
        options: { type: object, additionalProperties: true }
    CollectionInput:
      type: object
      required: [name, type, schema]
      properties:
        name: { type: string }
        type: { type: string, enum: [base, auth, view] }
        schema:
          type: array
          items: { $ref: "#/components/schemas/FieldSchema" }
        listRule: { type: string, nullable: true }
        viewRule: { type: string, nullable: true }
        createRule: { type: string, nullable: true }
        updateRule: { type: string, nullable: true }
        deleteRule: { type: string, nullable: true }
        authOptions: { type: object, additionalProperties: true }
    Collection:
      allOf:
        - $ref: "#/components/schemas/CollectionInput"
        - type: object
          properties:
            id: { type: string }
            created: { type: string }
            updated: { type: string }